src/requirements.txt
gspread==5.7.2
fastapi[all]==0.95.0
python_dotenv==0.21.0
# SQLAlchemy==1.4.44
# psycopg2-binary==2.9.6
